Move helpers to shared file
